refactor: 상품 api 스펙을 dto로 분리
