+++ src/main/resources/db/migration/V3__Alter_decimal_columns.sql
@@ -1,0 +1,2 @@
+ALTER TABLE menu ALTER COLUMN price BIGINT(20) NOT NULL;
+ALTER TABLE product ALTER COLUMN price BIGINT(20) NOT NULL;
